third_party/WebKit/LayoutTests/inspector-protocol/page/javascriptDialogEvents-expected.txt

CONFIRM NAVIGATION
ALERT: alert
CONFIRM: confirm
PROMPT: prompt, default text: 
Opening dialog: type=beforeunload; message=beforeunload in javascriptDialogEvents
Closed dialog: result=false
Opening dialog: type=alert; message=alert
Closed dialog: result=true
Opening dialog: type=confirm; message=confirm
Closed dialog: result=true
Opening dialog: type=prompt; message=prompt
Closed dialog: result=true
